[Yocto s5p6818] OPTEE Build Makefile modification
This patch is for the OPTEE Build Makefile modification.

When building Android Pie, the build structure was changed from
the existing script method to the lunch & make method.
There was a history of changing the path of the optee Make File
by multi-thread build.
so when Yocto Build, you have to apply this fix.

Signed-off-by: chungho.shin <chungho.shin@coasianexell.com>
Change-Id: Iaf60d169134d6d55993c7ec438806fd510ded7f8